Handle allow-multiple question tasks
