+++ app/models/reducers/stats_reducer.rb
@@ -3,7 +3,7 @@ module Reducers
     attr_reader :sub_ranges
 
     def reduction_data_for(extractions)
-      ReductionResults.build do |results|
+      CountingHash.build do |results|
         extractions.each do |extraction|
           extraction.data.each do |key, value|
             case value
